+++ benchmark/CMakeLists.txt
@@ -113,7 +113,6 @@ set(SWIFT_BENCH_MODULES
     single-source/ObserverUnappliedMethod
     single-source/OpaqueConsumingUsers
     single-source/OpenClose
-    single-source/PartialApplyDynamicType
     single-source/Phonebook
     single-source/PointerArithmetics
     single-source/PolymorphicCalls
